+++ CHANGELOG.md
@@ -10,7 +10,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### To Add
 
 * add usage of [transcript](https://www.tutorialspoint.com/how-to-use-a-transcript-in-powershell)
-* add progressbar or information like "[7 / 12] users in progress"
 * add translation files
 * added default value `$false` to each optional parameter `$beVerbose`
 * add user remove logic from [windows_remove_old_users](https://github.com/stevleibelt/windows_remove_old_users)
@@ -44,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+* Added display of progress bar if `$beVerbose = $false`
 * Added asking ":: Remove lock file (y/N)" if lock file exists
 * Added optional local host configuration `local-<hostname>-Configuration.ps1`
 * Added configuration value `$deleteRecycleBin` with default value `$false`